fix(client-generator-ts): start Deno config lookup from output dir
